Refactor menus.py into subdir
