+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/__init__.py
@@ -1,0 +1,4 @@
+from .ap import ApMenuPage
+from .carry_on import CarryOnMenuPage
+from .open_browser import OpenBrowserMenuPage
+from .welcome import WelcomeMenuPage
