+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/attr/margins.py
@@ -1,0 +1,5 @@
+# Formatting text in miniscreen
+INFO_PAGE_MARGIN_X = 29
+FIRST_LINE_Y = 9
+SECOND_LINE_Y = 25
+THIRD_LINE_Y = 41
