+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/attr/speeds.py
@@ -1,0 +1,6 @@
+# Tunings to approximately match other sys info pages using GIFs
+ANIMATION_SPEED = 1
+ANIMATION_SLEEP_INTERVAL = 0.005
+STATIONARY_SLEEP_INTERVAL = 0.5
+FIRST_DRAW_SLEEP_INTERVAL = 1
+DEFAULT_INTERVAL = 0.2
